Moved classes out of mask_subscription
